Move the Behaviors logic into the behaviors mixin
This involves a couple things:

1. extract parseBehaviors, BehaviorsTriggerBuilder, and behaviorsLookup
out of the Behaviors src file

2. move the Behaviors constructor logic to initBehaviors

3. finally kill the `Behaviors.wrap` function which was setting the
behaviorEvents function on the prototype. It's my most hated function.
