+++ src/utils/parseBehaviors.js
@@ -1,0 +1,37 @@
+import _                        from 'underscore';
+import _getValue                from '../utils/_getValue';
+
+// Takes care of getting the behavior class
+// given options and a key.
+// If a user passes in options.behaviorClass
+// default to using that.
+// If a user passes in a Behavior Class directly, use that
+// Otherwise delegate the lookup to the users `behaviorsLookup` implementation.
+function getBehaviorClass(options, key) {
+  if (options.behaviorClass) {
+    return options.behaviorClass;
+    //treat functions as a Behavior constructor
+  } else if (_.isFunction(options)) {
+    return options;
+  }
+
+  // behaviorsLookup can be either a flat object or a method
+  return _getValue(Marionette.Behaviors.behaviorsLookup, this, [options, key])[key];
+}
+
+// Iterate over the behaviors object, for each behavior
+// instantiate it and get its grouped behaviors.
+// This accepts a list of behaviors in either an object or array form
+function parseBehaviors(view, behaviors) {
+  return _.chain(behaviors).map(function(options, key) {
+    var BehaviorClass = getBehaviorClass(options, key);
+    //if we're passed a class directly instead of an object
+    var _options = options === BehaviorClass ? {} : options;
+    var behavior = new BehaviorClass(_options, view);
+    var nestedBehaviors = parseBehaviors(view, _.result(behavior, 'behaviors'));
+
+    return [behavior].concat(nestedBehaviors);
+  }).flatten().value();
+}
+
+export default parseBehaviors;
